+++ module-apps/application-notes/NotesModel.cpp
@@ -26,7 +26,7 @@ void NotesModel::requestRecords(const uint32_t offset, const uint32_t limit)
     DBServiceAPI::NotesGetLimitOffset(application, offset, limit);
 }
 
-bool NotesModel::updateRecords(std::unique_ptr<std::vector<NotesRecord>> records)
+bool NotesModel::updateRecords(std::vector<NotesRecord> records)
 {
 #if DEBUG_DB_MODEL_DATA == 1
     LOG_DEBUG("Offset: %" PRIu32 ", Limit: %" PRIu32 " Count: %" PRIu32 "", offset, limit, count);